Move toolbar logic to separate files
 - This makes it easier for devs to understand what is toolbar logic and what is not.
