+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/component/golden-layout/golden-layout.view.tsx
@@ -27,22 +27,15 @@ import user from '../singletons/user-instance'
 // @ts-expect-error ts-migrate(7016) FIXME: Could not find a declaration file for module 'sani... Remove this comment to see the full error message
 import sanitize from 'sanitize-html'
 import Button from '@mui/material/Button'
-import Grid from '@mui/material/Grid'
-import MinimizeIcon from '@mui/icons-material/Minimize'
-import FullscreenExitIcon from '@mui/icons-material/FullscreenExit'
-import FullscreenIcon from '@mui/icons-material/Fullscreen'
-import CloseIcon from '@mui/icons-material/Close'
 import ExtensionPoints from '../../extension-points/extension-points'
 import { Visualizations } from '../visualization/visualizations'
 import { LazyQueryResult } from '../../js/model/LazyQueryResult/LazyQueryResult'
 import { useListenTo } from '../selection-checkbox/useBackbone.hook'
 import Paper from '@mui/material/Paper'
-import MuiTooltip, { TooltipProps } from '@mui/material/Tooltip'
 import { Elevations } from '../theme/theme'
 import { useLazyResultsFromSelectionInterface } from '../selection-interface/hooks'
 import { LazyQueryResults } from '../../js/model/LazyQueryResult/LazyQueryResults'
 import { TypedUserInstance } from '../singletons/TypedUser'
-import PopoutIcon from '@mui/icons-material/OpenInNew'
 import { useHistory } from 'react-router-dom'
 import {
   Dialog,
@@ -50,25 +43,8 @@ import {
   DialogContent,
   DialogTitle,
 } from '@mui/material'
-
-/**
- *  There's a bit of funkiness with normal tooltip behavior when we click the minimize buttons, as the tooltip sticks behind and flickers.
- *  This ensures that when the user clicks the button, the tooltip will be hidden.
- */
-function Tooltip(props: TooltipProps) {
-  const [show, setShow] = React.useState(false)
-
-  return (
-    <MuiTooltip
-      open={show}
-      disableHoverListener
-      onMouseEnter={() => setShow(true)}
-      onMouseLeave={() => setShow(false)}
-      onMouseDown={() => setShow(false)}
-      {...props}
-    />
-  )
-}
+import { StackToolbar, HeaderHeight, MinimizedHeight } from './stack-toolbar'
+import { GoldenLayoutComponentHeader } from './visual-toolbar'
 
 const treeMap = (obj: any, fn: any, path = []): any => {
   if (Array.isArray(obj)) {
@@ -100,9 +76,6 @@ const sanitizeTree = (tree: any) =>
     }
     return obj
   })
-
-const HeaderHeight = 44
-const MinimizedHeight = HeaderHeight + 5
 
 function getGoldenLayoutSettings() {
   return {
@@ -138,125 +111,6 @@ const GoldenLayoutWindowCommunicationEvents = {
   consumeSubwindowLayoutChange: 'consumeSubwindowLayoutChange',
   consumeNavigationChange: 'consumeNavigationChange',
   consumeWreqrEvent: 'consumeWreqrEvent',
-}
-
-function useStackRelatedToTab(tab: GoldenLayout.Tab) {
-  const [stack, setStack] = React.useState(tab.contentItem.parent)
-  React.useEffect(() => {
-    setStack(tab.contentItem.parent)
-  }, [tab])
-  return stack as GoldenLayout.Tab & GoldenLayout.ContentItem
-}
-
-const GoldenLayoutComponentHeader = ({
-  viz,
-  tab,
-  options,
-  componentState,
-  container,
-  name,
-}: {
-  viz: any
-  tab: GoldenLayout.Tab & GoldenLayout.ContentItem
-  options: any
-  componentState: any
-  container: any
-  name: any
-}) => {
-  const relatedStack = useStackRelatedToTab(tab)
-  const { height, width } = useStackSize(relatedStack)
-  const { minimizeCallback } = usePixelHeightTrackingForTab(
-    tab as any,
-    relatedStack,
-    height
-  )
-  const canBeMinimized = useCanBeMinimizedTab({
-    stack: relatedStack,
-    width,
-    height,
-  })
-  const isMaximized = useIsMaximized({ stack: relatedStack })
-  const isMinimized = height && height <= MinimizedHeight
-  return (
-    <ExtensionPoints.providers>
-      <div
-        data-id={`${name}-tab`}
-        className={`flex flex-row items-center flex-nowrap`}
-      >
-        <Grid item className="px-2 text-lg">
-          <div>{tab.titleElement.text()}</div>
-        </Grid>
-        <Grid item>
-          {viz.header ? (
-            <viz.header
-              {..._.extend({}, options, componentState, {
-                container,
-              })}
-            />
-          ) : null}
-        </Grid>
-        {isMinimized || isMaximized || !canBeMinimized ? (
-          <></>
-        ) : (
-          <Grid item>
-            <Tooltip
-              title={
-                <Paper elevation={Elevations.overlays} className="p-2">
-                  Minimize visual to bottom of layout
-                </Paper>
-              }
-            >
-              <Button onClick={minimizeCallback}>
-                <MinimizeIcon />
-              </Button>
-            </Tooltip>
-          </Grid>
-        )}
-
-        <Grid item>
-          {!tab.contentItem.layoutManager.isSubWindow &&
-          tab.closeElement[0].style.display !== 'none' ? (
-            <Tooltip
-              title={
-                <Paper elevation={Elevations.overlays} className="p-2">
-                  Open visual in new window
-                </Paper>
-              }
-            >
-              <Button
-                data-id="popout-tab-button"
-                onClick={() => {
-                  tab.contentItem.popout()
-                }}
-              >
-                <PopoutIcon />
-              </Button>
-            </Tooltip>
-          ) : null}
-        </Grid>
-        <Grid item>
-          {tab.closeElement[0].style.display !== 'none' ? (
-            <Tooltip
-              title={
-                <Paper elevation={Elevations.overlays} className="p-2">
-                  Close visual
-                </Paper>
-              }
-            >
-              <Button
-                data-id="close-tab-button"
-                onClick={(e) => {
-                  ;(tab as any)._onCloseClickFn(e)
-                }}
-              >
-                <CloseIcon />
-              </Button>
-            </Tooltip>
-          ) : null}
-        </Grid>
-      </div>
-    </ExtensionPoints.providers>
-  )
 }
 
 /**
@@ -678,51 +532,13 @@ function handleGoldenLayoutStackCreated(stack: any) {
   let intervalId = setInterval(() => {
     try {
       const renderRoot = createRoot(stack.header.controlsContainer[0])
-      renderRoot.render(<GoldenLayoutToolbar stack={stack} />)
+      renderRoot.render(<StackToolbar stack={stack} />)
       stack.on('destroy', () => {
         renderRoot.unmount()
       })
       clearInterval(intervalId)
     } catch (err) {}
   }, 100)
-}
-
-/**
- *  Will return a safe, unchanging reference, feel free to keep in a variable
- * @param stack
- * @returns
- */
-function getRoot(
-  stack: GoldenLayout.ContentItem | GoldenLayout.Tab
-): GoldenLayout.ContentItem | undefined {
-  if (!stack) {
-    return undefined
-  }
-  const stackAsContentItem = stack as GoldenLayout.ContentItem
-  if (stackAsContentItem?.isRoot) {
-    return stackAsContentItem
-  }
-  let parent = stackAsContentItem.parent
-  if (parent?.type === 'root') {
-    return parent
-  } else {
-    return getRoot(parent)
-  }
-}
-
-/**
- *  This can change, so do not store it in a variable, instead keep around the Root from getRoot and refind the base item as necessary.
- * @param stack
- * @returns
- */
-function getRootColumnContent(
-  stack: GoldenLayout.ContentItem | GoldenLayout.Tab
-): GoldenLayout.ContentItem | undefined {
-  const root = getRoot(stack)
-  if (root) {
-    return root.contentItems[0]
-  }
-  return undefined
 }
 
 function useContainerSize(container: GoldenLayout.Container) {
@@ -745,544 +561,6 @@ function useContainerSize(container: GoldenLayout.Container) {
     return () => {}
   }, [container])
   return { height, width }
-}
-
-function useStackSize(stack: GoldenLayout.Tab & GoldenLayout.ContentItem) {
-  if (stack.titleElement) {
-    stack = stack.contentItem.parent as any // this is a tab, so find the stack related to the tab
-  }
-  const [width, setWidth] = React.useState<number | undefined>(
-    stack.header.element.width()
-  )
-  const [height, setHeight] = React.useState<number | undefined>(
-    getStackHeight({ stack })
-  )
-
-  React.useEffect(() => {
-    if (stack) {
-      const resizeCallback = () => {
-        setWidth(stack.header.element.width())
-        setHeight(getStackHeight({ stack }))
-      }
-      stack.on('resize', resizeCallback)
-      return () => {
-        stack.off('resize', resizeCallback)
-      }
-    }
-    return () => {}
-  }, [stack])
-  return { height, width }
-}
-
-// by ready, we mean the stack to be minimized is already on the bottom of a column layout => aka we can just resize it
-function layoutIsAlreadyReady({
-  stack,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-}) {
-  const bottomItem = getBottomItem(stack)
-  const isBottomItem = stack === bottomItem
-  return (
-    (isBottomItem && getRootColumnContent(stack)?.isColumn) ||
-    (!getRootColumnContent(stack)?.isColumn &&
-      !getRootColumnContent(stack)?.isRow)
-  )
-}
-
-// check if a minimized stack already exists
-function layoutAlreadyHasMinimizedStack({
-  stack,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-}) {
-  const bottomItem = getBottomItem(stack)
-
-  return (
-    getRootColumnContent(stack)?.isColumn &&
-    (bottomItem as any).pixelHeight <= MinimizedHeight
-  )
-}
-
-// add the stack to the existing minimized stack
-function addStackToExistingMinimizedStack({
-  stack,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-}) {
-  const bottomItem = getBottomItem(stack)
-  if (bottomItem) {
-    stack.contentItems.slice().forEach((thing) => {
-      stack.removeChild(thing as any, true) // for some reason removeChild is overly restrictive on type of "thing" so we have to cast
-      bottomItem.addChild(thing, undefined)
-    })
-  }
-}
-
-// add the tab to the existing minimized stack
-function addTabToExistingMinimizedStack({
-  tab,
-  stack,
-}: {
-  tab: GoldenLayout.Tab & GoldenLayout.ContentItem
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-}) {
-  const bottomItem = getBottomItem(stack)
-  if (bottomItem) {
-    stack.removeChild(tab.contentItem as any, true) // for some reason removeChild is overly restrictive on type of "thing" so we have to cast
-    bottomItem.addChild(tab.contentItem, undefined)
-  }
-}
-
-function rootIsNotAColumn(goldenLayoutRoot: GoldenLayout.ContentItem) {
-  return (
-    getRootColumnContent(goldenLayoutRoot) &&
-    !getRootColumnContent(goldenLayoutRoot)?.isColumn
-  )
-}
-
-function rootIsEmpty(goldenLayoutRoot: GoldenLayout.ContentItem) {
-  return !getRootColumnContent(goldenLayoutRoot)
-}
-
-// we have to move each item individually because golden layout doesn't support moving an entire stack, and addChild does not work as documentation says (it doesn't remove the existing automatically)
-function createNewStackFromExistingStack({
-  stack,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-}) {
-  const existingItems = stack.contentItems.slice()
-  const newStackItem = stack.layoutManager._$normalizeContentItem({
-    type: 'stack',
-  })
-  existingItems.forEach((thing) => {
-    stack.removeChild(thing as any, true) // for some reason removeChild is overly restrictive on type of "thing" so we have to cast
-    newStackItem.addChild(thing, undefined)
-  })
-  return newStackItem
-}
-
-// we have to move each item individually because golden layout doesn't support moving an entire stack, and addChild does not work as documentation says (it doesn't remove the existing automatically)
-function createNewStackFromExistingTab({
-  tab,
-  stack,
-}: {
-  tab: GoldenLayout.Tab & GoldenLayout.ContentItem
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-}) {
-  const existingItem = tab.contentItem
-  const newStackItem = stack.layoutManager._$normalizeContentItem({
-    type: 'stack',
-  })
-  stack.removeChild(existingItem as any, true) // for some reason removeChild is overly restrictive on type of "thing" so we have to cast
-  newStackItem.addChild(existingItem, undefined)
-  return newStackItem
-}
-
-// create a new minimized stack and add the stack to it
-function createAndAddNewMinimizedStack({
-  stack,
-  goldenLayoutRoot,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-  goldenLayoutRoot: GoldenLayout.ContentItem
-}) {
-  const newStackItem = createNewStackFromExistingStack({ stack })
-  if (rootIsNotAColumn(goldenLayoutRoot)) {
-    const existingRootContent = getRootColumnContent(goldenLayoutRoot)
-    goldenLayoutRoot.removeChild(existingRootContent as any, true) // for some reason removeChild is overly restrictive on type of "thing" so we have to cast
-
-    // we need a column for minimize to work, so make a new column and add the existing root to it
-    const newColumnItem = stack.layoutManager._$normalizeContentItem({
-      type: 'column',
-    })
-    newColumnItem.addChild(existingRootContent)
-    newColumnItem.addChild(newStackItem)
-
-    goldenLayoutRoot.addChild(newColumnItem)
-  } else if (rootIsEmpty(goldenLayoutRoot)) {
-    const newColumnItem = stack.layoutManager._$normalizeContentItem({
-      type: 'column',
-    })
-    newColumnItem.addChild(newStackItem)
-
-    goldenLayoutRoot.addChild(newColumnItem)
-  } else {
-    getRootColumnContent(goldenLayoutRoot)?.addChild(newStackItem)
-  }
-}
-
-function createAndAddNewMinimizedStackForTab({
-  tab,
-  stack,
-  goldenLayoutRoot,
-}: {
-  tab: GoldenLayout.Tab & GoldenLayout.ContentItem
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-  goldenLayoutRoot: GoldenLayout.ContentItem
-}) {
-  const newStackItem = createNewStackFromExistingTab({ stack, tab })
-  if (rootIsNotAColumn(goldenLayoutRoot)) {
-    const existingRootContent = getRootColumnContent(goldenLayoutRoot)
-    goldenLayoutRoot.removeChild(existingRootContent as any, true) // for some reason removeChild is overly restrictive on type of "thing" so we have to cast
-
-    // we need a column for minimize to work, so make a new column and add the existing root to it
-    const newColumnItem = stack.layoutManager._$normalizeContentItem({
-      type: 'column',
-    })
-    newColumnItem.addChild(existingRootContent)
-    newColumnItem.addChild(newStackItem)
-
-    goldenLayoutRoot.addChild(newColumnItem)
-  } else if (rootIsEmpty(goldenLayoutRoot)) {
-    const newColumnItem = stack.layoutManager._$normalizeContentItem({
-      type: 'column',
-    })
-    newColumnItem.addChild(newStackItem)
-
-    goldenLayoutRoot.addChild(newColumnItem)
-  } else {
-    getRootColumnContent(goldenLayoutRoot)?.addChild(newStackItem)
-  }
-}
-
-function getStackInMinimizedLocation({
-  goldenLayoutRoot,
-}: {
-  goldenLayoutRoot: GoldenLayout.ContentItem
-}): GoldenLayout.Container | undefined {
-  const rootColumnContent = getRootColumnContent(goldenLayoutRoot)
-  if (rootColumnContent) {
-    return (
-      rootColumnContent.contentItems[
-        rootColumnContent.contentItems.length - 1
-      ].getActiveContentItem() as any
-    ).container as GoldenLayout.Container
-  }
-  return undefined
-}
-
-// the true height of the stack - the golden layout implementation at the moment only tracks the height of the visual within the stack, not the stack itself
-function getTrueHeight({
-  goldenLayoutRoot,
-}: {
-  goldenLayoutRoot: GoldenLayout.ContentItem
-}) {
-  return (
-    getStackInMinimizedLocation({ goldenLayoutRoot })?.parent.parent
-      .element as unknown as GoldenLayout.Header['element']
-  ).height()
-}
-
-function adjustStackInMinimizedPlaceIfNecessary({
-  goldenLayoutRoot,
-}: {
-  goldenLayoutRoot: GoldenLayout.ContentItem
-}) {
-  if (getRootColumnContent(goldenLayoutRoot)?.isColumn) {
-    const trueHeight = getTrueHeight({ goldenLayoutRoot })
-    const minimizedStack = getStackInMinimizedLocation({ goldenLayoutRoot })
-    if (minimizedStack) {
-      minimizedStack.height = trueHeight || minimizedStack.height // otherwise setSize will not work correctly! - this allows us to consistently and always set the height to what we want!
-      minimizedStack.setSize(10, HeaderHeight)
-    }
-  }
-}
-
-// keep track of pixel height on each stack, which allows us to detect when a stack is "minimized" later on
-function usePixelHeightTracking(
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem,
-  height: number | undefined
-) {
-  React.useEffect(() => {
-    if (stack) {
-      ;(stack as any).pixelHeight = height
-    }
-  }, [height, stack])
-  const goldenLayoutRoot = useRoot(stack)
-
-  const minimizeCallback = React.useCallback(() => {
-    if (!goldenLayoutRoot) {
-      return
-    }
-    if (layoutIsAlreadyReady({ stack })) {
-      // do nothing? just resize to be minimized
-    } else if (layoutAlreadyHasMinimizedStack({ stack })) {
-      // minimized area exists, add this to it
-      addStackToExistingMinimizedStack({ stack })
-    } else {
-      // rearrange layout if necessary and move the stack
-      createAndAddNewMinimizedStack({ stack, goldenLayoutRoot })
-    }
-    adjustStackInMinimizedPlaceIfNecessary({ goldenLayoutRoot })
-  }, [stack, goldenLayoutRoot])
-
-  return { minimizeCallback }
-}
-
-// keep track of pixel height on each stack, which allows us to detect when a stack is "minimized" later on
-function usePixelHeightTrackingForTab(
-  tab: GoldenLayout.Tab & GoldenLayout.ContentItem,
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem,
-  height: number | undefined
-) {
-  React.useEffect(() => {
-    if (stack) {
-      ;(stack as any).pixelHeight = height
-    }
-  }, [height, stack])
-  const goldenLayoutRoot = useRoot(stack)
-
-  const minimizeCallback = React.useCallback(() => {
-    if (!goldenLayoutRoot) {
-      return
-    }
-    if (layoutAlreadyHasMinimizedStack({ stack })) {
-      // minimized area exists, add this to it
-      addTabToExistingMinimizedStack({ tab, stack })
-    } else {
-      // rearrange layout if necessary and move the stack
-      createAndAddNewMinimizedStackForTab({ tab, stack, goldenLayoutRoot })
-    }
-    adjustStackInMinimizedPlaceIfNecessary({ goldenLayoutRoot })
-  }, [stack, goldenLayoutRoot])
-
-  return { minimizeCallback }
-}
-
-function useRoot(stack: GoldenLayout.Tab & GoldenLayout.ContentItem) {
-  const [root, setRoot] = React.useState(getRoot(stack))
-  React.useEffect(() => {
-    setRoot(getRoot(stack))
-  }, [stack])
-  return root
-}
-
-function getBottomItem(
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-): GoldenLayout.ContentItem | undefined {
-  const rootColumnContent = getRootColumnContent(stack)
-  if (rootColumnContent) {
-    return rootColumnContent.contentItems[
-      rootColumnContent.contentItems.length - 1
-    ]
-  }
-  return undefined
-}
-
-function useCanBeMinimized({
-  stack,
-  height,
-  width,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-  height?: number
-  width?: number
-}) {
-  const [canBeMinimized, setCanBeMinimized] = React.useState(true)
-  React.useEffect(() => {
-    const rootContent = getRootColumnContent(stack)
-    if (rootContent?.isStack) {
-      setCanBeMinimized(false)
-    } else {
-      setCanBeMinimized(true)
-    }
-  }, [stack, height, width])
-  return canBeMinimized
-}
-
-function useCanBeMinimizedTab({
-  stack,
-  height,
-  width,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-  height?: number
-  width?: number
-}) {
-  const [itemCount, setItemCount] = React.useState(stack.contentItems.length)
-  const [canBeMinimized, setCanBeMinimized] = React.useState(true)
-  React.useEffect(() => {
-    const rootContent = getRootColumnContent(stack)
-    if (rootContent?.isStack && rootContent?.contentItems.length === 1) {
-      setCanBeMinimized(false)
-    } else {
-      setCanBeMinimized(true)
-    }
-  }, [stack, height, width, itemCount])
-  React.useEffect(() => {
-    if (stack) {
-      const callback = () => {
-        setItemCount(stack.contentItems.length)
-      }
-      stack.on('itemCreated', callback)
-      stack.on('itemDestroyed', callback)
-      return () => {
-        stack.off('itemCreated', callback)
-        stack.off('itemDestroyed', callback)
-      }
-    }
-    return () => {}
-  }, [stack])
-  return canBeMinimized
-}
-
-function useIsMaximized({
-  stack,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-}) {
-  const [isMaximized, setIsMaximized] = React.useState(false)
-  React.useEffect(() => {
-    const maximizedCallback = () => {
-      setIsMaximized(true)
-    }
-    stack.on('maximised', maximizedCallback)
-    const minimizedCallback = () => {
-      setIsMaximized(false)
-    }
-    stack.on('minimised', minimizedCallback)
-    return () => {
-      stack.off('maximised', maximizedCallback)
-      stack.off('minimised', minimizedCallback)
-    }
-  }, [stack])
-  React.useEffect(() => {
-    setIsMaximized(stack.isMaximised)
-  }, [stack])
-  return isMaximized
-}
-
-// to avoid fixing types everywhere, let's make a function
-function getStackHeight({
-  stack,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-}) {
-  return (stack.element as unknown as GoldenLayout.Header['element']).height()
-}
-
-const GoldenLayoutToolbar = ({
-  stack,
-}: {
-  stack: GoldenLayout.Tab & GoldenLayout.ContentItem
-}) => {
-  const isMaximized = useIsMaximized({ stack })
-  const { height, width } = useStackSize(stack)
-  const { minimizeCallback } = usePixelHeightTracking(stack, height)
-  const canBeMinimized = useCanBeMinimized({ stack, width, height })
-  const isMinimized = height && height <= MinimizedHeight
-  return (
-    <ExtensionPoints.providers>
-      <div className="flex flex-row flex-nowrap items-center">
-        <>
-          {isMaximized || isMinimized || !canBeMinimized ? (
-            <></>
-          ) : (
-            <div>
-              <Tooltip
-                title={
-                  <Paper elevation={Elevations.overlays} className="p-2">
-                    Minimize stack of visuals to bottom of layout
-                  </Paper>
-                }
-              >
-                <Button
-                  data-id="minimise-layout-button"
-                  onClick={minimizeCallback}
-                >
-                  <MinimizeIcon />
-                </Button>
-              </Tooltip>
-            </div>
-          )}
-          {isMaximized ? (
-            <div>
-              <Tooltip
-                title={
-                  <Paper elevation={Elevations.overlays} className="p-2">
-                    Restore stack of visuals to original size
-                  </Paper>
-                }
-              >
-                <Button
-                  data-id="unmaximise-layout-button"
-                  onClick={() => {
-                    stack.toggleMaximise()
-                  }}
-                >
-                  <FullscreenExitIcon />
-                </Button>
-              </Tooltip>
-            </div>
-          ) : (
-            <div>
-              <Tooltip
-                title={
-                  <Paper elevation={Elevations.overlays} className="p-2">
-                    Maximize stack of visuals
-                  </Paper>
-                }
-              >
-                <Button
-                  data-id="maximise-layout-button"
-                  onClick={() => {
-                    stack.toggleMaximise()
-                  }}
-                >
-                  <FullscreenIcon />
-                </Button>
-              </Tooltip>
-            </div>
-          )}
-
-          {stack.layoutManager.isSubWindow ? null : (
-            <div>
-              <Tooltip
-                title={
-                  <Paper elevation={Elevations.overlays} className="p-2">
-                    Open stack of visuals in new window
-                  </Paper>
-                }
-              >
-                <Button
-                  data-id="popout-layout-button"
-                  onClick={() => {
-                    stack.popout()
-                  }}
-                >
-                  <PopoutIcon fontSize="small" />
-                </Button>
-              </Tooltip>
-            </div>
-          )}
-
-          <div>
-            {(stack.header as any)._isClosable() ? (
-              <Tooltip
-                title={
-                  <Paper elevation={Elevations.overlays} className="p-2">
-                    Close stack of visuals
-                  </Paper>
-                }
-              >
-                <Button
-                  data-id="close-layout-button"
-                  onClick={() => {
-                    if (stack.isMaximised) {
-                      stack.toggleMaximise()
-                    }
-                    stack.remove()
-                  }}
-                >
-                  <CloseIcon />
-                </Button>
-              </Tooltip>
-            ) : null}
-          </div>
-        </>
-      </div>
-    </ExtensionPoints.providers>
-  )
 }
 
 const useUpdateGoldenLayoutSize = ({
